Stop using uwsgi due to dev environment
uwsgi is not working in a development environment as containers
are making non-secure (i.e. http) requests to other containers.

In two places we want to make requests to core django to verify user
info:
* from cms container: https://github.com/TACC/Core-CMS/blob/main/taccsite_cms/remote_cms_auth.py
* from protx container's onboarded_user_required decorator

We want to it be unsecure as requesting directly from container in dev
environment (as not able to currently make a request cep.dev/localhost from
container)