Update pyproject.toml python and numpy versions (#31)
* Update pyproject.toml

As of numpy 1.21.2, there is a maximum version of python. Numpy 1.21.2 (to 1.21.5 as time of writing this answer) requires python version greater than 3.7 but strictly lower than 3.11.

Also, Python 3.7 is no longer supported in release 1.22.0 and later of numpy. Python support has been dropped. This is rather strict, there are changes that require Python >= 3.8.

This edit should allow us to resolve the numpy version conflict. 
Also, bumped numpy up to the next-to-latest release (which has a deprecation in it and I didn't want to trigger a side-effect by accident).

* Updated the python version ranges to support the update to numpy. Also bumped the gevent version as it was the pinned below important updates.

* Had to pin the werkzeug version and drop back flask a small bit due to a version constraint on python version.